+++ BroadleafCommerceIntegrationTests/src/test/java/org/broadleafcommerce/content/ContentDetailsDaoProvider.java
@@ -28,7 +28,7 @@ public class ContentDetailsDaoProvider {
 	@DataProvider(name = "basicContentDetails")
 	public static Object[][] provideBasicContentDetails() {
 		ContentDetails cd = new ContentDetailsImpl();
-		cd.setId(Long.getLong("1919"));
+		cd.setId(Integer.getInteger("1919"));
 		cd.setXmlContent("<xml>some xml content</xml>");
 		return new Object[][] {{cd}};
 	}
